+++ CHANGELOG.md
@@ -49,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Remove support for the ESP8266 (#576)
 - Remove the direct boot image format (#577)
+- Remove support for Raspberry Pi's internal UART peripherals (#585)
 
 ## [2.1.0] - 2023-10-03